# app/server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.MenuButtonItem:
    properties:
      buttonCode:
        type: string
      buttonDesc:
        type: string
      id:
        type: integer
    type: object
  dto.MenuButtonRequest:
    properties:
      buttonCode:
        maxLength: 64
        type: string
      buttonDesc:
        type: string
      menuId:
        type: integer
    required:
    - buttonCode
    - menuId
    type: object
...
  dto.MenuNode:
    properties:
      activeMenu:
        type: string
      buttons:
        items:
          $ref: '#/definitions/model.SysMenuButton'
        type: array
      children:
        items:
          $ref: '#/definitions/dto.MenuNode'
        type: array
      component:
        type: string
      constant:
        type: boolean
      createBy:
        type: integer
      createTime:
        type: string
      fixedIndexInTab:
        type: integer
      hideInMenu:
        type: boolean
      href:
        type: string
      i18nKey:
        type: string
      icon:
        type: string
      iconType:
        $ref: '#/definitions/model.IconType'
      id:
        type: integer
      isSystem:
        type: boolean
      keepAlive:
        type: boolean
      menuName:
        type: string
      menuType:
        $ref: '#/definitions/model.MenuType'
      multiTab:
        type: boolean
      order:
        type: integer
      parentId:
        type: integer
      query:
        $ref: '#/definitions/model.JSONMap'
      routeName:
        type: string
      routePath:
        type: string
      status:
        $ref: '#/definitions/model.EnableStatus'
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.MenuRequest:
    properties:
      activeMenu:
        type: string
      buttons:
        items:
          $ref: '#/definitions/dto.MenuButtonItem'
        type: array
      component:
        type: string
      constant:
        type: boolean
      fixedIndexInTab:
        type: integer
      hideInMenu:
        type: boolean
      href:
        type: string
      i18nKey:
        type: string
      icon:
        type: string
      iconType:
        $ref: '#/definitions/model.IconType'
      keepAlive:
        type: boolean
      menuName:
        maxLength: 64
        type: string
      menuType:
        allOf:
        - $ref: '#/definitions/model.MenuType'
        enum:
        - "1"
        - "2"
      multiTab:
        type: boolean
      order:
        type: integer
      parentId:
        type: integer
      query: {}
      routeName:
        maxLength: 128
        type: string
      routePath:
        maxLength: 255
        type: string
      status:
        $ref: '#/definitions/model.EnableStatus'
    required:
    - menuName
    - menuType
    - routeName
    - routePath
    type: object
...
  dto.MenuSearch:
    properties:
      current:
        type: integer
      keyword:
        type: string
      menuName:
        type: string
      size:
        type: integer
      status:
        $ref: '#/definitions/model.EnableStatus'
    type: object
...
  dto.PageResponse:
    properties:
      current:
        type: integer
      records: {}
      size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.TagRequest:
    properties:
      description:
        type: string
      tagName:
        maxLength: 64
        type: string
    required:
    - tagName
    type: object
  dto.TagSearch:
    properties:
      current:
        type: integer
      keyword:
        type: string
      size:
        type: integer
      tagName:
        type: string
    type: object
...
  model.BookTag:
    properties:
      createBy:
        type: integer
      createTime:
        type: string
      description:
        type: string
      id:
        type: integer
      tagName:
        type: string
      updateBy:
        type: integer
      updateTime:
        type: string
      usageCount:
        type: integer
    type: object
...
  model.EnableStatus:
    enum:
    - "1"
    - "2"
    type: string
    x-enum-varnames:
    - StatusEnabled
    - StatusDisabled
...
  model.IconType:
    enum:
    - "1"
    - "2"
    type: string
    x-enum-varnames:
    - IconTypeIconify
    - IconTypeLocal
  model.JSONMap:
    additionalProperties: {}
    type: object
...
  model.MenuType:
    enum:
    - "1"
    - "2"
    type: string
    x-enum-comments:
      MenuTypeDir: 目录
      MenuTypeMenu: 菜单
    x-enum-descriptions:
    - 目录
    - 菜单
    x-enum-varnames:
    - MenuTypeDir
    - MenuTypeMenu
...
  model.SysMenu:
    properties:
      activeMenu:
        type: string
      component:
        type: string
      constant:
        type: boolean
      createBy:
        type: integer
      createTime:
        type: string
      fixedIndexInTab:
        type: integer
      hideInMenu:
        type: boolean
      href:
        type: string
      i18nKey:
        type: string
      icon:
        type: string
      iconType:
        $ref: '#/definitions/model.IconType'
      id:
        type: integer
      isSystem:
        type: boolean
      keepAlive:
        type: boolean
      menuName:
        type: string
      menuType:
        $ref: '#/definitions/model.MenuType'
      multiTab:
        type: boolean
      order:
        type: integer
      parentId:
        type: integer
      query:
        $ref: '#/definitions/model.JSONMap'
      routeName:
        type: string
      routePath:
        type: string
      status:
        $ref: '#/definitions/model.EnableStatus'
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  model.SysMenuButton:
    properties:
      buttonCode:
        type: string
      buttonDesc:
        type: string
      createTime:
        type: string
      id:
        type: integer
      menuId:
        type: integer
      updateTime:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 小说阅读平台后端 API
  title: Boread API
  version: "1.0"
paths:
...
  /api/book/tag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BookTag'
              type: object
      security:
      - BearerAuth: []
      summary: 新增标签
      tags:
      - book-tag
  /api/book/tag/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - book-tag
    get:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BookTag'
              type: object
      security:
      - BearerAuth: []
      summary: 标签详情
      tags:
      - book-tag
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BookTag'
              type: object
      security:
      - BearerAuth: []
      summary: 编辑标签
      tags:
      - book-tag
  /api/book/tag/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TagSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 标签分页
      tags:
      - book-tag
...
  /api/manage/menu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SysMenu'
              type: object
      security:
      - BearerAuth: []
      summary: 新增菜单
      tags:
      - menu
  /api/manage/menu/{id}:
    delete:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - menu
    get:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SysMenu'
              type: object
      security:
      - BearerAuth: []
      summary: 菜单详情
      tags:
      - menu
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SysMenu'
              type: object
      security:
      - BearerAuth: []
      summary: 编辑菜单
      tags:
      - menu
  /api/manage/menu/button:
    post:
      consumes:
      - application/json
      parameters:
      - description: 按钮
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuButtonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SysMenuButton'
              type: object
      security:
      - BearerAuth: []
      summary: 新增菜单按钮
      tags:
      - menu
  /api/manage/menu/button/{id}:
    delete:
      parameters:
      - description: 按钮ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单按钮
      tags:
      - menu
  /api/manage/menu/buttons/{menuId}:
    get:
      parameters:
      - description: 菜单ID
        in: path
        name: menuId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SysMenuButton'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 按菜单查按钮
      tags:
      - menu
  /api/manage/menu/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuNode'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 菜单分页列表 (平级, 含按钮)
      tags:
      - menu
  /api/manage/menu/tree:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuNode'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 全量菜单树 (含按钮)
      tags:
      - menu
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"